MAINTAINERS: rectify radix-tree testing entry in XARRAY
The initial commit 3d5bd6e1a04a ("xarray: Add MAINTAINERS entry")
missed a trailing slash to include all files and subdirectory files.
Hence, all files in tools/testing/radix-tree were not part of XARRAY,
but were considered to be part of "THE REST".

Rectify this to ensure patches reach the actual maintainer.

This was identified with a small script that finds all files only
belonging to "THE REST" according to the current MAINTAINERS file, and I
acted upon its output.

Fixes: 3d5bd6e1a04a ("xarray: Add MAINTAINERS entry")
Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>