MAINTAINERS: add headers and doc to SCHEDULER
Various files in include/linux/sched/ are identified as part of THE REST
according to MAINTAINERS, but they are really belong to SCHEDULER.

Add include/linux/sched/, include/uapi/linux/sched/ and
Documentation/scheduler/ to the SCHEDULER entry, and order the entry with
parse-maintainers.pl.

This was identified with a small script that finds all files belonging to
THE REST according to the current MAINTAINERS file, and I investigated
upon its output.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>