MAINTAINERS: include GOOGLE FIRMWARE entry
All files in drivers/firmware/google/ are identified as part of THE REST
according to MAINTAINERS, but they are really maintained by others.

Add a basic entry for drivers/firmware/google/ based on a simple statistics
on tags of commits in that directory:

  $ git log drivers/firmware/google/ | grep '\-by:' \
      | sort | uniq -c | sort -nr
     62     Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
     13     Reviewed-by: Guenter Roeck <groeck@chromium.org>
     12     Signed-off-by: Stephen Boyd <swboyd@chromium.org>
     11     Reviewed-by: Julius Werner <jwerner@chromium.org>

There is no specific mailing list for this driver, based on observations
on the patch emails, and the git history suggests the driver is maintained.

This was identified with a small script that finds all files belonging to
THE REST according to the current MAINTAINERS file, and I investigated
upon its output.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>