MAINTAINERS: clarify maintenance of Gemini SoC driver
Commit 68198dca4569 ("soc: Add SoC driver for Gemini") introduced a Gemini
SoC driver, but did not add the driver to the existing ARM/CORTINA SYSTEMS
GEMINI ARM ARCHITECTURE entry in MAINTAINERS.

Hence, this driver was considered to be part of "THE REST".

Clarify now that this driver is maintained by the ARM/CORTINA SYSTEMS
GEMINI ARM ARCHITECTURE maintainers.

This was identified with a small script that finds all files only
belonging to "THE REST" according to the current MAINTAINERS file, and I
acted upon its output.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>