lib: add static_key-based error injection framework
This new error-injection framework uses static_keys for the definition of an error injection point. If enabled, these error injection points allows to replace each expression with an (expected) error value.

To define a errorkey-injection point, each expression, which may cause the error, can be instrumented. The error value, which can be injected, is defined by the errorkey type.

If an injected errorkey is disable, the instrumentation just executes the instrumented expression.

If the errorkey is enabled, the according errorkey_filter function is called. These filter functions will replace the execution of the instrumented expression with the selected error value. It is intended, to add filter definitions, which allows error injections on enabled errorkeys only on matching filter criteria.

Errorkey injection example I:

  ptr = kmalloc(sizeof(*ptr), GFP_KERNEL);
  if (!ptr)
      /* error handling */

kmalloc() returns NULL in case of an error. So instrument the calling expression with a NULL-pointer errorkey injection.

  ptr = ERROR_NULL(kmalloc(sizeof(*ptr), GFP_KERNEL));
  if (!ptr)
      /* error handling */

Errorkey injection example II:

  if(s->val < 0)
      /* error handling */

In this case, a value below 0 in member val of structure s is erroneous. Instrument that case with a TRUE errorkey injection.

  if(ERROR_TRUE(s->val < 0)
      /* error handling */

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Holger Dengler <dengler@linutronix.de>

Signed-off-by: Saurabh Arora <saurabh.arora@elektrobit.com>