arm: pgtable: refer to intended CONFIG_ARM_LPAE in comment
Commit 5615f69bc209 ("ARM: 9016/2: Initialize the mapping of KASan shadow
memory") adds some nested ifdef's in ./arch/arm/mm/pgd.c, and follows the
good practice to annotate the endif's with a comment to indicate the
corresponding ifdef condition.

One comment annotation refers to CONFIG_LPAE, whereas the config is
actually called CONFIG_ARM_LPAE. That imprecision in a comment is probably
tolerable for all human readers.

However, the script ./scripts/checkkconfigsymbols.py, which checks the
kernel tree for references to non-existing Kconfig symbols, identifies and
reports that the reference to CONFIG_LPAE is invalid.

The script ./scripts/checkkconfigsymbols.py has been quite useful to
identify a number of bugs with Kconfig symbols and deserves to be executed
and checked regularly.

So, repair the comment to reduce the reports from this script and simplify
to use this script, as new issues are easier to spot when the list of
reports is shorter.

Signed-off-by: Lukas Bulwahn <lukas.bulwahn@gmail.com>